LPS-102660 Additional information when deleting an Experience with a running AB Test
