+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/content/Language.properties
@@ -27,6 +27,7 @@ content-value=Content Value
 contents=Contents
 create-new-content=Create New Content
 delete-experience=Delete Experience
+delete-experience-with-running-test-confirmation-message=There is an existent test for this experience. When deleting the Experience all tests inside will be deleted too. Are you sure you want to delete this Experience and Tests inside?
 deprioritize-experience=Deprioritize Experience
 discard-variant=Discard Variant
 do-you-want-to-delete-this-experience=Do you want to delete this experience? Its content will be deleted.
